Feature calendar widget (#384)
* XCXC - Stuff I needed to get running

* Fix reference error in Content Feed button handling

* Ignore emacs backup files

* Add CalendarWidget model

* Copy directory_widgets views into calendar_views

* Add CalendarWidget route, views, and controller

This adds the CalendarWidget routes for both the embed rendering
as well as the business tools administration pages. The calendar
widgets are functional, in that you can create and edit them.
However, they do not render appropriately and their new/edit forms
are not formatted desirably.

* Order form elements, Add TODO for Appcues

* Add Calendar Feeds to the Website Builder

This adds the calendar feeds to the website builder. You can add
a calendar feed on any page that can have a SupportLocal embed,
and it follows similar configuration behavior.

Notably missing are the icon for the calendar (and styles for that
modal). This also renders the block very similarly to a normal
content feed at the moment.

* Clean up FeedContentModal

* Grossly position the calendar list view styles

There is a lot of cleanup to do here, but this commit takes the
view template and styles it in broad strokes so that it is
basically in the right layout.

* Clean up event grid item styles

* Styles for calendar feed list view

* Basic agenda styles (with some xcxc hacks to remove)

* Some hacky date tabs for agenda

* Some basic collapsing search fields

* Switch between calendar feed views based on param

* Add default_view to calendar_widgets

This adds a default_view option to calendar widgets, so that admins
can select whether to display the feed as a list, grid, or agenda.
If search is not enabled for the widget, this is the only view the
feed will use.

* Use the same calendar rendering for native feeds and iframes

* Hook up ChangeView buttons in calendar feed

* Create search function for calendar widget w date filter

* Agenda highlights given day and returns results for that day

* Working, if ugly, agenda date tabs

* Fix up agenda styles

* Prevent caching when new calendar feed params are present

* Use existing date picker for calendar search

* Select week view by default for new calendar embeds

* Add messages for calendar embeds

* Use List view for searches

* Use separate param for agenda tab controls

* Return correct events for all views

* Add event total_counts

* Fix indentation issue with placeholder images

* Tidy a little

* Add Bootstrap panel classes

* Style agenda nav with Bootstrap columns

* Use bootstrap rows and cols uniformly

* Use Bootstrap styles in event search bar

* Get grid view working nicely

* Fix agenda view styles

* Fix up remaining styles

* Move files together in website directory

* Remove caching in website groups

* Fix sidebar activeness in nested menus

* Small style fixes

* More style overrides of Bootstrap style overrides

* Expand search when params exist, plus more style fixes

* Updates label

* Removes extra margin top