Prevent duplicate content creation (#415)
* Migrate TLD from .dev to .test

* Prevent duplicate QuickPosts

This fixes two bugs. The first involved TurboLinks duplicating JS
listeners. While it seems poor to be including the JavaScript twice,
the quick fix is to ensure any JavaScript is idempotent. Using .off()
I ensure that the code can be run as many times as needed without
duplicating effects.

The second patch reworks the JavaScript listeners a bit so that the
submit buttons are properly disabled upon click.

Other events are likely to be susceptible to similar bugs.

* Prevent duplicates in other content types