Summernote text editor updates (#421)
* Remove Air mode in RichTextEditors

This disables Summernote's air mode for the RichTextEditor
component. Instead, it uses the default toolbar setting and
hides/shows the toolbar based on focusin and focusout events.

* WIP partial summernote update

* WIP summernote update with cleaner

* Save changes to state on update

* Hide Summernote handles in web blocks

* Use Summernote React component in QuickPosts

* Match editor styles to theme in webpage blocks

RichTextEditors not have transparent backgrounds with a border on
hover, so that the webpage blocks as closely match the webpage
theme as possible. When focused, the editor has a translucent white
background to highlight the editable area without completely
obscuring the background (or messing with contrast of custom font
colors).

* Update summernote-cleaner

* Use RichTextEditor for all content types

This updates content types to use the RichTextEditor component
powered by Summernote. These forms already used Summernote through
a custom jQuery initialization, so this change primarily unifies
them around existing behavior.

Placeholder text, which as far as I could tell wasn't being shown
anyways, was removed from the code in this commit, as it caused a
problem with the slideUp/slideDown behavior.

* Remove class attributes on clean

* Plus icon when empty on hover

* Move cleaner paintbrush icon down

* Remove placeholder code when pasting to empty summernote