Will summernote tuning (#461)
* Strip more tags from inline editors

* Add RichTextEditor to contact form description

* Highlight active header style in RichText dropdown

* Prefer HTML formatted paste text over plain

There is browser-specific behavior in what type of data is
available when you paste. If the pasted text is coming directly
from an HTML source, everything is happy. If the source is rich
text, some browsers convert that to HTML (Chrome, Safari), while
others only have plain text available (Firefox). To maintain cross
browser compatibility, we try to take either one, but prefer HTML
if it is there.

* Remove even more tags from inline editors

* Add style for active summernote buttons

* Use RichTextEditor in Todo comments